requirements.txt
supervision
python-dotenv